CivilEngineeringInsightStudio/requirements.txt
streamlit
python-dotenv
Pillow
google-generativeai==0.3.2
